Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/of-config-verify-beryllium-mvn33-openjdk7/3/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-verify-beryllium-mvn33-openjdk7/3/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-verify-beryllium-mvn33-openjdk7/3

https://jenkins.opendaylight.org/releng/job/of-config-verify-beryllium-mvn33-openjdk8/3/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-verify-beryllium-mvn33-openjdk8/3/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-verify-beryllium-mvn33-openjdk8/3

https://jenkins.opendaylight.org/releng/job/of-config-distribution-check-beryllium/2/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-distribution-check-beryllium/2/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-distribution-check-beryllium/2

https://jenkins.opendaylight.org/releng/job/of-config-validate-autorelease-beryllium/1/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-validate-autorelease-beryllium/1/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-validate-autorelease-beryllium/1

Patch-set: 1