Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/of-config-verify-boron-mvn33-openjdk8/10/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-verify-boron-mvn33-openjdk8/10/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-verify-boron-mvn33-openjdk8/10

https://jenkins.opendaylight.org/releng/job/of-config-validate-autorelease-boron/7/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-validate-autorelease-boron/7/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-validate-autorelease-boron/7

https://jenkins.opendaylight.org/releng/job/of-config-distribution-check-boron/8/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/of-config-distribution-check-boron/8/ )

Logs: https://logs.opendaylight.org/releng/jenkins092/of-config-distribution-check-boron/8

Patch-set: 1