Update patch set 3
Patch Set 3:

ok of-config committers - this patch has been open for a while - needs movement here.  of-config has signed up for Oxygen so it's not dead. At some point soon the CSS will be removed and this project won't compile/run so you'll have to review/merge this in order to move forward or be left in the dust. So let's just get it done....

Patch-set: 3