Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/of-config-validate-autorelease-oxygen/14/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng/job/of-config-validate-autorelease-oxygen/14/ )

Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/of-config-validate-autorelease-oxygen/14

https://jenkins.opendaylight.org/releng/job/of-config-distribution-check-oxygen/17/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng/job/of-config-distribution-check-oxygen/17/ )

Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/of-config-distribution-check-oxygen/17

https://jenkins.opendaylight.org/releng/job/of-config-maven-verify-oxygen-mvn33-openjdk8/10/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng/job/of-config-maven-verify-oxygen-mvn33-openjdk8/10/ )

Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/of-config-maven-verify-oxygen-mvn33-openjdk8/10

Patch-set: 1