Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultaneous Release. Add a profile that is active by
default so that default behavior is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I706254ea60621f07e6a4fa6c0f7de22d9ce2a2aa
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>