Update patch set 2
Patch Set 2: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/of-config-validate-autorelease-nitrogen/15/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng/job/of-config-validate-autorelease-nitrogen/15/ )

Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/of-config-validate-autorelease-nitrogen/15

https://jenkins.opendaylight.org/releng/job/of-config-maven-verify-nitrogen-mvn33-openjdk8/5/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng/job/of-config-maven-verify-nitrogen-mvn33-openjdk8/5/ )

Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/of-config-maven-verify-nitrogen-mvn33-openjdk8/5

https://jenkins.opendaylight.org/releng/job/of-config-distribution-check-nitrogen/20/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng/job/of-config-distribution-check-nitrogen/20/ )

Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/of-config-distribution-check-nitrogen/20

Patch-set: 2
Tag: ci